+++ api/src/main/kotlin/io/nightfish/lightnovelreader/api/ui/ReaderStyle.kt
@@ -1,0 +1,11 @@
+package io.nightfish.lightnovelreader.api.ui
+
+import androidx.compose.ui.graphics.Color
+
+data class ReaderStyle(
+    val fontSize: Float,
+    val fontLineHeight: Float,
+    val fontWeight: Float,
+    val textColor: Color,
+    val textDarkColor: Color
+)
