+++ app/src/main/kotlin/indi/dmzz_yyhyy/lightnovelreader/MainActivity.kt
@@ -8,9 +8,12 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
+import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.intl.Locale
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
@@ -29,10 +32,12 @@ import indi.dmzz_yyhyy.lightnovelreader.theme.LightNovelReaderTheme
 import indi.dmzz_yyhyy.lightnovelreader.ui.LightNovelReaderApp
 import indi.dmzz_yyhyy.lightnovelreader.utils.LogUtils
 import io.nightfish.lightnovelreader.api.bookshelf.BookshelfSortType
+import io.nightfish.lightnovelreader.api.ui.ReaderStyle
 import io.nightfish.lightnovelreader.api.userdata.UserDataPath
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.cancel
+import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.launch
 import java.util.concurrent.TimeUnit
 import javax.inject.Inject
@@ -122,7 +127,36 @@ class MainActivity : ComponentActivity() {
             }
         }
 
+        val fontSizeUserData = userDataRepository.floatUserData(UserDataPath.Reader.FontSize.path)
+        val fontLineHeightUserData = userDataRepository.floatUserData(UserDataPath.Reader.FontLineHeight.path)
+        val fontWeightUserData = userDataRepository.floatUserData(UserDataPath.Reader.FontWeigh.path)
+        val textColorUserData = userDataRepository.colorUserData(UserDataPath.Reader.TextColor.path)
+        val textDarkColorUserData = userDataRepository.colorUserData(UserDataPath.Reader.TextDarkColor.path)
+        val fontFamilyUriUserData = userDataRepository.uriUserData(UserDataPath.Reader.FontFamilyUri.path)
         setContent {
+            val readerStyle by remember {
+                combine(
+                    fontSizeUserData.getFlowWithDefault(15f),
+                    fontLineHeightUserData.getFlowWithDefault(7f),
+                    fontWeightUserData.getFlowWithDefault(500f),
+                    textColorUserData.getFlowWithDefault(Color.Unspecified),
+                    textDarkColorUserData.getFlowWithDefault(Color.Unspecified)
+                ) { fontSize, lineHeight, weight, textColor, textDarkColor ->
+                    ReaderStyle(
+                        fontSize = fontSize,
+                        fontLineHeight = lineHeight,
+                        fontWeight = weight,
+                        textColor = textColor,
+                        textDarkColor = textDarkColor,
+                    )
+                }
+            }.collectAsState(initial = ReaderStyle(
+                fontSize = 15f,
+                fontLineHeight = 7f,
+                fontWeight = 500f,
+                textColor = Color.Unspecified,
+                textDarkColor = Color.Unspecified,
+            ))
             LightNovelReaderTheme(
                 darkMode = darkMode,
                 appLocale = appLocale,
@@ -132,6 +166,7 @@ class MainActivity : ComponentActivity() {
                 darkThemeName = darkThemeName
             ) {
                 LightNovelReaderApp(
+                    readerStyle = readerStyle,
                     onBuildNavHost = {
                         with(pluginManager) {
                             onBuildNavHost()
