+++ app/src/main/kotlin/indi/dmzz_yyhyy/lightnovelreader/ui/book/reader/content/componet/SimpleTextComponent.kt
@@ -35,6 +35,15 @@ fun SimpleTextComponentContent(
             color = color,
             lineHeight = (fontSize.value + fontLineHeight.value).sp,
             onTextLayout = {
+                println(
+                    fontSize.toString() + "\n" +
+                    fontLineHeight.toString() + "\n" +
+                    fontWeight.toString() + "\n"+
+                    fontFamily.toString() + "\n"+
+                    color.toString() + "\n" +
+                    text.slice(0..10) + "\n" +
+                    "ciallo!"
+                )
                 println("ciallo")
             }
         )
