+++ app/src/main/kotlin/indi/dmzz_yyhyy/lightnovelreader/ui/navigation/NavHost.kt
@@ -45,12 +45,15 @@ import indi.dmzz_yyhyy.lightnovelreader.utils.expandExit
 import indi.dmzz_yyhyy.lightnovelreader.utils.expandPopEnter
 import indi.dmzz_yyhyy.lightnovelreader.utils.expandPopExit
 import io.nightfish.lightnovelreader.api.ui.LocalNavController
+import io.nightfish.lightnovelreader.api.ui.LocalReaderStyle
+import io.nightfish.lightnovelreader.api.ui.ReaderStyle
 
 @OptIn(ExperimentalMaterial3Api::class, ExperimentalSharedTransitionApi::class)
 @Composable
 fun LightNovelReaderNavHost(
     navController: NavHostController,
-    onBuildNavHost: NavGraphBuilder.() -> Unit
+    onBuildNavHost: NavGraphBuilder.() -> Unit,
+    readerStyle: ReaderStyle
 ) {
     val snackbarHostState = remember { SnackbarHostState() }
 
@@ -58,8 +61,8 @@ fun LightNovelReaderNavHost(
     val claim: (Boolean) -> Unit = remember { { take -> claimCount += if (take) 1 else -1 } }
 
     var bottomBarVisible by remember { mutableStateOf(true) }
-
     CompositionLocalProvider(
+        LocalReaderStyle provides readerStyle,
         LocalNavController provides navController,
         LocalSnackbarHost provides snackbarHostState,
         LocalClaimSnackbarHost provides claim,
